Merges BruisePackTrigger and OintmentTrigger into MedicalStackTrigger Now these two items will have a defined amount of uses, wont heal limbs that are healthy and characters who try to heal themselves will have a progress action. Healing others is still instantaneous. Bleeding speed is greatly reduced and will start at 20 brute damage on a single limb. Several limbs bleeding at the same time will increase the amount of blood loss. Limb bleeding stops when the brute damage of said limb is below 20. Fixes Attack() using the server's character bodylimb aim instead of the originator.

+++ UnityProject/Assets/Scripts/Health/Living/LivingHealthBehaviour.cs
@@ -532,7 +532,7 @@ public abstract class LivingHealthBehaviour : NetworkBehaviour
 	/// <summary>
 	/// Updates the blood health stats from the server via NetMsg
 	/// </summary>
-	public void UpdateClientBloodStats(int heartRate, int bloodVolume, float oxygenDamage, float toxinLevel)
+	public void UpdateClientBloodStats(int heartRate, float bloodVolume, float oxygenDamage, float toxinLevel)
 	{
 		bloodSystem.UpdateClientBloodStats(heartRate, bloodVolume, oxygenDamage, toxinLevel);
 	}
